+++ internal/brain/capturing_test.go
@@ -9,47 +9,47 @@ import (
 
 func TestCheckCapture(t *testing.T) {
 	// Initialize
-	Game.Goban.Tab = [board.TOT_SIZE]int{}
+	Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	Game.CurrentPlayer = Game.P1
-	center := (board.SIZE * board.SIZE) / 2
+	center := int16((board.SIZE * board.SIZE) / 2)
 	if board.SIZE%2 == 0 {
 		center += board.SIZE / 2
 	}
 	tables := []struct {
-		position                  int
-		opponentPositions         []int
-		currentPlayerPositions    []int
-		expectedCaptureDirections []int
+		position                  int16
+		opponentPositions         []int16
+		currentPlayerPositions    []int16
+		expectedCaptureDirections []int16
 	}{
 		{
 			position:                  center,
-			opponentPositions:         []int{},
-			currentPlayerPositions:    []int{},
-			expectedCaptureDirections: []int{},
+			opponentPositions:         []int16{},
+			currentPlayerPositions:    []int16{},
+			expectedCaptureDirections: []int16{},
 		},
 		{
 			position:                  center,
-			opponentPositions:         []int{center + 1, center + 2},
-			currentPlayerPositions:    []int{center, center + 3},
-			expectedCaptureDirections: []int{E},
+			opponentPositions:         []int16{center + 1, center + 2},
+			currentPlayerPositions:    []int16{center, center + 3},
+			expectedCaptureDirections: []int16{E},
 		},
 		{
 			position:                  center,
-			opponentPositions:         []int{center - 1, center - 2},
-			currentPlayerPositions:    []int{center, center - 3},
-			expectedCaptureDirections: []int{W},
+			opponentPositions:         []int16{center - 1, center - 2},
+			currentPlayerPositions:    []int16{center, center - 3},
+			expectedCaptureDirections: []int16{W},
 		},
 		{
 			position:                  center,
-			opponentPositions:         []int{center - board.SIZE, center - (2 * board.SIZE)},
-			currentPlayerPositions:    []int{center, center - (3 * board.SIZE)},
-			expectedCaptureDirections: []int{N},
+			opponentPositions:         []int16{center - board.SIZE, center - (2 * board.SIZE)},
+			currentPlayerPositions:    []int16{center, center - (3 * board.SIZE)},
+			expectedCaptureDirections: []int16{N},
 		},
 		{
 			position:                  center,
-			opponentPositions:         []int{center + board.SIZE, center + (2 * board.SIZE)},
-			currentPlayerPositions:    []int{center, center + (3 * board.SIZE)},
-			expectedCaptureDirections: []int{S},
+			opponentPositions:         []int16{center + board.SIZE, center + (2 * board.SIZE)},
+			currentPlayerPositions:    []int16{center, center + (3 * board.SIZE)},
+			expectedCaptureDirections: []int16{S},
 		},
 	}
 	for _, table := range tables {
@@ -65,14 +65,14 @@ func TestCheckCapture(t *testing.T) {
 			t.Errorf("Wrong captureDirections for %d, expected %v, got %v", table.position, table.expectedCaptureDirections, captureDirections)
 
 		}
-		Game.Goban.Tab = [board.TOT_SIZE]int{}
+		Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	}
 }
 
 func BenchmarkCheckCapture(b *testing.B) {
-	Game.Goban.Tab = [board.TOT_SIZE]int{}
+	Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	Game.CurrentPlayer = Game.P1
 	for i := 0; i < b.N; i++ {
-		checkCapture(i%(board.SIZE*board.SIZE), &Game.Goban.Tab, Game.CurrentPlayer.Id)
+		checkCapture(int16(i)%(board.SIZE*board.SIZE), &Game.Goban.Tab, Game.CurrentPlayer.Id)
 	}
 }
