+++ internal/brain/sequences_test.go
@@ -9,47 +9,47 @@ import (
 
 func TestCheckSequence(t *testing.T) {
 	// Initialize
-	Game.Goban.Tab = [board.TOT_SIZE]int{}
+	Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	Game.CurrentPlayer = Game.P1
-	center := (board.SIZE * board.SIZE) / 2
+	center := int16((board.SIZE * board.SIZE) / 2)
 	if board.SIZE%2 == 0 {
 		center += board.SIZE / 2
 	}
 	tables := []struct {
-		position               int
-		opponentPositions      []int
-		currentPlayerPositions []int
-		expectedSequences      []int
+		position               int16
+		opponentPositions      []int16
+		currentPlayerPositions []int16
+		expectedSequences      []int16
 	}{
 		{
 			position:               center,
-			opponentPositions:      []int{},
-			currentPlayerPositions: []int{},
-			expectedSequences:      []int{0, 0, 0, 0, 0, 0, 0, 0},
+			opponentPositions:      []int16{},
+			currentPlayerPositions: []int16{},
+			expectedSequences:      []int16{0, 0, 0, 0, 0, 0, 0, 0},
 		},
 		{
 			position:               center,
-			opponentPositions:      []int{center + 1, center + 2},
-			currentPlayerPositions: []int{center, center + 3},
-			expectedSequences:      []int{0, 0, 0, 0, 0, 0, 0, 0},
+			opponentPositions:      []int16{center + 1, center + 2},
+			currentPlayerPositions: []int16{center, center + 3},
+			expectedSequences:      []int16{0, 0, 0, 0, 0, 0, 0, 0},
 		},
 		{
 			position:               center,
-			opponentPositions:      []int{center - 1, center - 2},
-			currentPlayerPositions: []int{center, center + 1, center - 3},
-			expectedSequences:      []int{0, 0, 1, 0, 0, 0, 0, 0},
+			opponentPositions:      []int16{center - 1, center - 2},
+			currentPlayerPositions: []int16{center, center + 1, center - 3},
+			expectedSequences:      []int16{0, 0, 1, 0, 0, 0, 0, 0},
 		},
 		{
 			position:               center,
-			opponentPositions:      []int{center - board.SIZE, center - (2 * board.SIZE)},
-			currentPlayerPositions: []int{center, center + board.SIZE, center - (board.SIZE), center + 1, center - 1, center - 2, center - 3},
-			expectedSequences:      []int{1, 1, 1, 3, 0, 0, 0, 0},
+			opponentPositions:      []int16{center - board.SIZE, center - (2 * board.SIZE)},
+			currentPlayerPositions: []int16{center, center + board.SIZE, center - (board.SIZE), center + 1, center - 1, center - 2, center - 3},
+			expectedSequences:      []int16{1, 1, 1, 3, 0, 0, 0, 0},
 		},
 		{
 			position:               center,
-			opponentPositions:      []int{},
-			currentPlayerPositions: []int{center, center - (board.SIZE) + 1, center - (board.SIZE) - 1, center + (board.SIZE) + 1, center + (board.SIZE) - 1},
-			expectedSequences:      []int{0, 0, 0, 0, 1, 1, 1, 1},
+			opponentPositions:      []int16{},
+			currentPlayerPositions: []int16{center, center - (board.SIZE) + 1, center - (board.SIZE) - 1, center + (board.SIZE) + 1, center + (board.SIZE) - 1},
+			expectedSequences:      []int16{0, 0, 0, 0, 1, 1, 1, 1},
 		},
 	}
 	for _, table := range tables {
@@ -66,51 +66,51 @@ func TestCheckSequence(t *testing.T) {
 			t.Errorf("Wrong sequenceLengths for %d, expected %v, got %v", table.position, table.expectedSequences, sequenceLengths)
 
 		}
-		// Game.Gban.Tab = make([]int, board.SIZE*board.SIZE)
-		Game.Goban.Tab = [board.TOT_SIZE]int{}
+		// Game.Gban.Tab = make([]int16, board.SIZE*board.SIZE)
+		Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	}
 }
 
 func BenchmarkCheckSequence(b *testing.B) {
-	Game.Goban.Tab = [board.TOT_SIZE]int{}
+	Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	Game.CurrentPlayer = Game.P1
 	for i := 0; i < b.N; i++ {
-		CheckSequence(i%(board.SIZE*board.SIZE), 1, &Game.Goban.Tab)
+		CheckSequence(int16(i)%(board.SIZE*board.SIZE), 1, &Game.Goban.Tab)
 	}
 }
 
 func TestCompleteSequenceForPosition(t *testing.T) {
 	// Initialize
-	Game.Goban.Tab = [board.TOT_SIZE]int{}
+	Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	Game.CurrentPlayer = Game.P1
-	center := (board.SIZE * board.SIZE) / 2
+	center := int16((board.SIZE * board.SIZE) / 2)
 	if board.SIZE%2 == 0 {
 		center += board.SIZE / 2
 	}
 	tables := []struct {
-		position               int
-		currentPlayerPositions []int
-		expectedSequences      [][]int
+		position               int16
+		currentPlayerPositions []int16
+		expectedSequences      [][]int16
 	}{
 		{
 			position:               center,
-			currentPlayerPositions: []int{center, center + 1, center - 1},
-			expectedSequences:      [][]int{[]int{center, center - 1, center + 1}},
+			currentPlayerPositions: []int16{center, center + 1, center - 1},
+			expectedSequences:      [][]int16{[]int16{center, center - 1, center + 1}},
 		},
 		{
 			position:               center,
-			currentPlayerPositions: []int{center, center + board.SIZE, center - board.SIZE},
-			expectedSequences:      [][]int{[]int{center, center - board.SIZE, center + board.SIZE}},
+			currentPlayerPositions: []int16{center, center + board.SIZE, center - board.SIZE},
+			expectedSequences:      [][]int16{[]int16{center, center - board.SIZE, center + board.SIZE}},
 		},
 		{
 			position:               center,
-			currentPlayerPositions: []int{center, center + (board.SIZE + 1), center - (board.SIZE + 1)},
-			expectedSequences:      [][]int{[]int{center, center - (board.SIZE + 1), center + (board.SIZE + 1)}},
+			currentPlayerPositions: []int16{center, center + (board.SIZE + 1), center - (board.SIZE + 1)},
+			expectedSequences:      [][]int16{[]int16{center, center - (board.SIZE + 1), center + (board.SIZE + 1)}},
 		},
 		{
 			position:               center,
-			currentPlayerPositions: []int{center, center + (board.SIZE - 1), center - (board.SIZE - 1)},
-			expectedSequences:      [][]int{[]int{center, center - (board.SIZE - 1), center + (board.SIZE - 1)}},
+			currentPlayerPositions: []int16{center, center + (board.SIZE - 1), center - (board.SIZE - 1)},
+			expectedSequences:      [][]int16{[]int16{center, center - (board.SIZE - 1), center + (board.SIZE - 1)}},
 		},
 	}
 	for _, table := range tables {
@@ -121,6 +121,6 @@ func TestCompleteSequenceForPosition(t *testing.T) {
 		if !reflect.DeepEqual(table.expectedSequences, completeSequences) {
 			t.Errorf("Wrong completeSequences for %d, expected %v, got %v", table.position, table.expectedSequences, completeSequences)
 		}
-		Game.Goban.Tab = [board.TOT_SIZE]int{}
+		Game.Goban.Tab = [board.TOT_SIZE]int16{}
 	}
 }
